get rid of save note use case and move everything to edit note use case
